Extract `Event::Organizable` module
The module has been extracted in order to reduce code duplication.
